Vectorization
Instead of looping over values, checks are performed by vectorized
pandas operations.

Some overhead is added to expand the qc flags string to separate columns
and then colapse them back into one column again. For ease of use, this
is done before and after each category of qc tests. This can
be changed in the future if extra speed is needed.

The class interfaces potentially have som residues from the earlier
implementation which was more object based.

Part of #32